refactor: 리팩토링 및 컴포넌트 경로 정리, 불필요한 props 제거
