+++ package.json
@@ -18,6 +18,7 @@
   "dependencies": {
     "@lukemorales/query-key-factory": "^1.3.4",
     "@radix-ui/react-dialog": "^1.1.14",
+    "@radix-ui/react-hover-card": "^1.1.14",
     "@tanstack/react-query": "^5.81.5",
     "@tanstack/react-query-devtools": "^5.81.5",
     "@vanilla-extract/css": "^1.17.4",
