+++ src/constants/routes.ts
@@ -9,7 +9,7 @@ export const PATH = {
   GAME_SELECT: '/game-select',
   GAME_AUTO_MATCHING: '/game-select/auto-matching',
   GAME_CUSTOM_MATCHING: '/game-select/custom-matching',
-  TOURNAMENT: '/game-select/tournament',
+  TOURNAMENT: '/game/tournament',
 
   FRIEND: '/friend',
   FRIEND_CHATROOM: '/friend/chatroom',
