+++ src/pages/GamePage/TournamentPage/components/MatchLines/styles.css.ts
