+++ src/pages/GamePage/TournamentPage/components/PlayerCard/styles.css.ts
