+++ src/pages/GamePage/TournamentPage/index.tsx
@@ -1,20 +1,18 @@
 import { useState, useEffect, useMemo } from 'react';
 import { useNavigate, useSearchParams } from 'react-router-dom';
 
-import type { ChatMessage, TournamentRoundType } from '@/api/types';
+import type { TournamentRoundType } from '@/api/types';
 import { BackButton } from '@/components/ui';
 
 import { roundTwoData } from './__mocks__/round2';
 import { roundFourData } from './__mocks__/round4';
-import { ChatBox, ChatInput } from './_components/Chat';
-import { MatchNode } from './_components/MatchNode';
+import { MatchNode } from './components/MatchNode';
 import * as styles from './styles.css';
 
 export const TournamentPage = () => {
   const navigate = useNavigate();
   const [searchParams] = useSearchParams();
 
-  const [messages, setMessages] = useState<ChatMessage[]>([]);
   const [readyIds, setReadyIds] = useState<string[]>([]);
 
   const currentUserId = 'pong'; // TODO: 실제 사용자 연동 (테스트용 임시 유저)
@@ -46,44 +44,21 @@ export const TournamentPage = () => {
 
   const tournamentData = round === 'ROUND_2' ? roundTwoData : roundFourData;
 
-  const handleSend = (msg: ChatMessage) => {
-    setMessages((prev) => [...prev, msg]);
-  };
-
   const handleReady = () => {
     if (!readyIds.includes(currentUserId)) {
       setReadyIds((prev) => [...prev, currentUserId]);
     }
   };
 
-  const renderDevToggleButtons = () => (
-    <div style={{ position: 'absolute', top: 4, right: 10 }}>
-      <button
-        style={{ padding: '4px 8px', fontSize: '13px' }}
-        onClick={() => navigate('/game/tournament?round=ROUND_2')}
-      >
-        Round 2 보기
-      </button>
-      <button
-        style={{ padding: '4px 8px', fontSize: '13px' }}
-        onClick={() => navigate('/game/tournament?round=ROUND_4')}
-      >
-        Round 4 보기
-      </button>
-    </div>
-  );
-
   return (
-    <div className={styles.container}>
-      {import.meta.env.MODE === 'development' && renderDevToggleButtons()}
-
+    <>
       <BackButton />
 
-      <div className={styles.bracketSection}>
-        <div className={styles.titleWrapper}>
+      <div className={styles.container}>
+        <header className={styles.header}>
           <h1 className={styles.title}>Tournament</h1>
           <h2 className={styles.round}>{round === 'ROUND_2' ? 'Round 2' : 'Round 4'}</h2>
-        </div>
+        </header>
 
         <MatchNode
           match={tournamentData}
@@ -92,15 +67,10 @@ export const TournamentPage = () => {
           currentUserId={currentUserId}
         />
 
-        <div className={styles.readyButtonWrapper}>
+        <div className={styles.buttonWrapper}>
           <button className={styles.readyButton} onClick={handleReady} aria-label="Ready" />
         </div>
       </div>
-
-      <div className={styles.chatSection}>
-        <ChatBox messages={messages} />
-        <ChatInput onSend={handleSend} />
-      </div>
-    </div>
+    </>
   );
 };
